requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-simplejwt
dj-database-url
django-rest-swagger
pyyaml